requirements.txt
openai
python-dotenv
networkx
matplotlib
river
pyyaml
